# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: rogier@codegorilla.nl
  description: Dutch Student Housing version 1
  title: Dutch Student Housing
  version: 0.1.0
...
servers:
- url: /
...
paths:
  /properties:
    get:
      description: Get properties with optional filters
      operationId: getProperties
      parameters:
      - description: City where the property is
        example: Groningen
        explode: false
        in: query
        name: city
        required: false
        schema:
          type: string
        style: form
      - description: Latitude
        explode: false
        in: query
        name: lat
        required: false
        schema:
          format: float
          type: number
        style: form
      - description: Longitude
        explode: false
        in: query
        name: long
        required: false
        schema:
          format: float
          type: number
        style: form
      - description: Distance in kilometers from known point
        explode: false
        in: query
        name: distance
        required: false
        schema:
          format: float
          type: number
        style: form
      - allowEmptyValue: true
        description: Items per page
        explode: false
        in: query
        name: perPage
        required: false
        schema:
          format: int32
          type: integer
        style: form
      - allowEmptyValue: true
        description: Page number (allows computed range on server side along with
          perPage parameter)
        explode: false
        in: query
        name: page
        required: false
        schema:
          format: int32
          type: integer
        style: form
      - description: Value to order listed items by
        explode: false
        in: query
        name: orderBy
        required: false
        schema:
          default: price
          enum:
          - price
          - pricepersqm
          type: string
        style: form
      - description: Minimum price
        explode: false
        in: query
        name: min
        required: false
        schema:
          format: float
          type: number
        style: form
      - description: Maximum price
        explode: false
        in: query
        name: max
        required: false
        schema:
          format: float
          type: number
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Properties'
            text/csv:
              schema:
                type: string
          description: Successful operation
        "400":
          description: Invalid paramters
        "404":
          description: No properties found for given parameters
        "401":
          description: Unauthorized to view properties (not yet logged in?)
      security:
      - dsh_auth:
        - read:properties
      summary: Get properties with optional filters
      tags:
      - property
      x-accepts: application/json
      x-tags:
      - tag: property
  /property:
    post:
      description: Add a new property listing
      operationId: addProperty
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Property'
        description: Create a new property listing
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
            text/csv:
              schema:
                type: string
          description: Successful operation
        "405":
          description: Invalid input
      security:
      - dsh_auth:
        - write:properties
        - read:properties
      summary: Add a new property listing
      tags:
      - property
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: property
    put:
      description: Update an existing property by Id
      operationId: updateProperty
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Property'
        description: Update an existent property listing
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
            text/csv:
              schema:
                type: string
          description: Successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Property not found
        "405":
          description: Validation exception
      security:
      - dsh_auth:
        - write:properties
        - read:properties
      summary: Update an existing property
      tags:
      - property
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: property
...
components:
  schemas:
    Property:
      example:
        img: img
        city: city
        date_published: 2000-01-23T04:56:07.000+00:00
        name: name
        id: id
        rent: 1.4658129
        postal_code: postal_code
        type: Room
        lat: 0.8008282
        long: 6.0274563
        sqm: 5
        rent_incl: true
      properties:
        id:
          type: string
        name:
          type: string
        city:
          type: string
        lat:
          format: float
          type: number
        long:
          format: float
          type: number
        img:
          type: string
        date_published:
          format: date-time
          type: string
        rent:
          format: float
          type: number
        rent_incl:
          type: boolean
        sqm:
          format: int32
          type: integer
        postal_code:
          type: string
        type:
          default: Room
          enum:
          - Room
          - Apartment
          - Studio
          - Other
          type: string
      type: object
    Properties:
      items:
        $ref: '#/components/schemas/Property'
      type: array
...
  securitySchemes:
    dsh_auth:
      bearerFormat: JWT
      scheme: bearer
      type: http